+++ tests/clients/test_publisher.py
@@ -5,6 +5,7 @@ import pika
 import pytest
 
 from yapw.clients import Base, Blocking, Durable, Transient
+from yapw.util import json_dumps
 
 
 class DurableClient(Durable, Blocking, Base):
@@ -15,6 +16,10 @@ class TransientClient(Transient, Blocking, Base):
     pass
 
 
+def dumps(message):
+    return b"overridden"
+
+
 @pytest.mark.parametrize("client_class", [DurableClient, TransientClient])
 @patch("pika.BlockingConnection")
 def test_init_default(connection, client_class):
@@ -23,17 +28,27 @@ def test_init_default(connection, client_class):
     client.channel.exchange_declare.assert_not_called()
 
     assert client.exchange == ""
+    assert client.encoder == json_dumps
+    assert client.content_type == "application/json"
     assert client.format_routing_key("test") == "_test"
 
 
 @pytest.mark.parametrize("client_class,durable", [(DurableClient, True), (TransientClient, False)])
 @patch("pika.BlockingConnection")
 def test_init_kwargs(connection, client_class, durable):
-    client = client_class(exchange="exch", exchange_type="fanout", routing_key_template="{routing_key}_{exchange}")
+    client = client_class(
+        exchange="exch",
+        exchange_type="fanout",
+        encoder=dumps,
+        content_type="application/octet-stream",
+        routing_key_template="{routing_key}_{exchange}",
+    )
 
     client.channel.exchange_declare.assert_called_once_with(exchange="exch", exchange_type="fanout", durable=durable)
 
     assert client.exchange == "exch"
+    assert client.encoder == dumps
+    assert client.content_type == "application/octet-stream"
     assert client.format_routing_key("test") == "test_exch"